pyproject.toml
[project]
name = "dev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dataset>=1.6.2",
    "huggingface-hub[cli]>=0.29.2",
    "llama-cpp-python>=0.3.8",
    "numpy>=2.2.5",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
]
